moved conncomp from laplace_beltrami.py to gen_isosurface.py
